requirements.txt
torch>=2.0
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
tqdm>=4.65
pyyaml>=6.0
joblib>=1.3
